Commit md-table, to output Markdown table
There's a lot of overlap with `gridify --format pipe` here.

The one thing that gridify (which uses tabulate) is lacking is the
ability to concisely override the column alignments, becuase tabulate
has all the control over that.

The one thing that md-table is lacking is all the fixed-width padding
and aligning that tabulate can do.

Argh.