Add table2json
usage: table2json [SELECTOR] [options...]

If SELECTOR is not provided, the default is simply "table".

All options at https://github.com/lightswitch05/table-to-json#options are
supported, EXCEPT for "ignoreHiddenRows", because something about jsdom
causes jQuery to think all rows are invisible. Also, the "extractor" option
takes CoffeeScript instead of JavaScript. For example:

    --extractor '(i, $cell) -> $cell.text().toUpperCase()'

Additionally, use "--local" to provide the directory of a local repo of
table-to-json to use your own version of the plugin.