Fix log() parsing
Although both `log stream ...` and `log show ...` list the same set of
column headers, only `log show ...` actually outputs a column for "TTL".
`log stream ...` does NOT output anything in this column, so we have to
check for whether it exists or not.

Also changed parsing to show whether the message is "Info", "Error", or
any other string.

Also changed the parsing to be smarter about how it strips junk words
from the text on the right-side of the log message.